# packages/mtmai/mtmai-0.3.798.tar.gz/mtmai-0.3.798/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.699"
  },
  "paths": {
...
    "/api/v1/admin/build_docker": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Build Docker",
        "operationId": "admin-build_docker",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/setup_database": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Setup Database",
        "operationId": "admin-setup_database",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/docker_build_base": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Docker Build Base",
        "operationId": "admin-docker_build_base",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/testing_fs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Testing Fs",
        "operationId": "admin-testing_fs",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/envs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Envs",
        "operationId": "admin-envs",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/usage": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Usage",
        "operationId": "admin-usage",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/git_clone_example": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Git Clone Example",
        "operationId": "admin-git_clone_example",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/dist_usage": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Dist Usage",
        "operationId": "admin-dist_usage",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/bash": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Run Bash",
        "operationId": "admin-run_bash",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBashReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/read_file": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Read File",
        "operationId": "admin-read_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadFileReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReadFileReq": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          }
        },
        "type": "object",
        "required": [
          "file_path"
        ],
        "title": "ReadFileReq"
      },
      "RunBashReq": {
        "properties": {
          "cmd": {
            "type": "string",
            "title": "Cmd"
          }
        },
        "type": "object",
        "required": [
          "cmd"
        ],
        "title": "RunBashReq"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}